+++ xcloud-devops-ci/xcloud-devops-ci-core/src/main/java/com/wl4g/devops/ci/core/param/RollbackParameter.java
@@ -29,7 +29,7 @@ public class RollbackParameter extends GenericParameter {
 		super();
 	}
 
-	public RollbackParameter(Integer pipeId, String remark) {
+	public RollbackParameter(Long pipeId, String remark) {
 		super(pipeId, remark);
 	}
